--- DEMEN/Assets/ui/prefab UI/GameManager.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 7359216185bd4bf46a63187c2e08ff97
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
