--- DEMEN/Assets/ui/scrip ui/PauseManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f743b442e55683d46a8e5c488641aac9
